MicrogridUP refactor into a sane python module
Hi guys, to complete the production deployment of MicrogridUP, I've refactored the project in this commit into a standard directory structure.

This was done to help Flask enforce its file access security and to help other users of the code understand what we're doing. Also, hopefully, it will make thing easier to find.

Major changes:
1. All templates now live in /templates.
2. All model runs now live in /data/projects/.
3. All test files now live in /testfiles.
4. The new file creation html + javascript + css was all merged into /templates/template_new.html
5. Inlined main.css into template_new.html
6. Updated various microgridup_*.py files to use new paths.
7. Updated .gitignore to preserve a couple projects in /data/projects but ignore others, so we have some demo ones backed into the repo but we don't push a ton of test data up there.